+++ compiler-plugin/src/main/kotlin/tech/mappie/ir/MappieContext.kt
@@ -35,51 +35,51 @@ fun allMappieClasses(): Set<IrClassSymbol> = setOf(
 
 context(context: MappieContext)
 fun referenceMappieClass(): IrClassSymbol =
-    context.pluginContext.referenceClass(CLASS_ID_MAPPIE)!!
+    context.pluginContext.finderForBuiltins().findClass(CLASS_ID_MAPPIE)!!
 
 context(context: MappieContext)
 fun referenceObjectMappieClass(): IrClassSymbol =
-    context.pluginContext.referenceClass(CLASS_ID_OBJECT_MAPPIE)!!
+    context.pluginContext.finderForBuiltins().findClass(CLASS_ID_OBJECT_MAPPIE)!!
 
 context(context: MappieContext)
 fun referenceObjectMappieClass2(): IrClassSymbol =
-    context.pluginContext.referenceClass(CLASS_ID_OBJECT_MAPPIE2)!!
+    context.pluginContext.finderForBuiltins().findClass(CLASS_ID_OBJECT_MAPPIE2)!!
 
 context(context: MappieContext)
 fun referenceObjectMappieClass3(): IrClassSymbol =
-    context.pluginContext.referenceClass(CLASS_ID_OBJECT_MAPPIE3)!!
+    context.pluginContext.finderForBuiltins().findClass(CLASS_ID_OBJECT_MAPPIE3)!!
 
 context(context: MappieContext)
 fun referenceObjectMappieClass4(): IrClassSymbol =
-    context.pluginContext.referenceClass(CLASS_ID_OBJECT_MAPPIE4)!!
+    context.pluginContext.finderForBuiltins().findClass(CLASS_ID_OBJECT_MAPPIE4)!!
 
 context(context: MappieContext)
 fun referenceObjectMappieClass5(): IrClassSymbol =
-    context.pluginContext.referenceClass(CLASS_ID_OBJECT_MAPPIE5)!!
+    context.pluginContext.finderForBuiltins().findClass(CLASS_ID_OBJECT_MAPPIE5)!!
 
 context(context: MappieContext)
 fun referenceEnumMappieClass(): IrClassSymbol =
-    context.pluginContext.referenceClass(CLASS_ID_ENUM_MAPPIE)!!
+    context.pluginContext.finderForBuiltins().findClass(CLASS_ID_ENUM_MAPPIE)!!
 
 context(context: MappieContext)
 fun referenceFunctionLet() =
-    context.pluginContext.referenceFunctions(CallableId(PACKAGE_KOTLIN, identifier("let"))).first()
+    context.pluginContext.finderForBuiltins().findFunctions(CallableId(PACKAGE_KOTLIN, identifier("let"))).first()
 
 context(context: MappieContext)
 fun referenceFunctionRequireNotNull() =
-    context.pluginContext.referenceFunctions(CallableId(PACKAGE_KOTLIN, identifier("requireNotNull"))).first {
+    context.pluginContext.finderForBuiltins().findFunctions(CallableId(PACKAGE_KOTLIN, identifier("requireNotNull"))).first {
         it.owner.hasShape(regularParameters = 2)
     }
 
 context(context: MappieContext)
 fun referenceFunctionRun() =
-    context.pluginContext.referenceFunctions(CallableId(PACKAGE_KOTLIN, identifier("run"))).first {
+    context.pluginContext.finderForBuiltins().findFunctions(CallableId(PACKAGE_KOTLIN, identifier("run"))).first {
         it.owner.hasShape(regularParameters = 1)
     }
 
 context(context: MappieContext)
 fun referenceFunctionError() =
-    context.pluginContext.referenceFunctions(CallableId(PACKAGE_KOTLIN, identifier("error"))).first()
+    context.pluginContext.finderForBuiltins().findFunctions(CallableId(PACKAGE_KOTLIN, identifier("error"))).first()
 
 context(context: MappieContext)
 fun IrEnumEntry.referenceFunctionValueOf(): IrSimpleFunction =